[ui] Remove old LogsScrollingTable
