+++ js_modules/dagster-ui/packages/ui-core/src/runs/RunFragments.tsx
@@ -1,6 +1,6 @@
 import {gql} from '@apollo/client';
 
-import {LOGS_SCROLLING_TABLE_MESSAGE_FRAGMENT} from './LogsScrollingTable';
+import {LOGS_SCROLLING_TABLE_MESSAGE_FRAGMENT} from './LogsScrollingTableMessageFragment';
 import {RUN_METADATA_PROVIDER_MESSAGE_FRAGMENT} from './RunMetadataProvider';
 import {RUN_TIMING_FRAGMENT} from './RunTimingDetails';
 import {EXECUTION_PLAN_TO_GRAPH_FRAGMENT} from '../gantt/toGraphQueryItems';
