+++ js_modules/dagster-ui/packages/ui-core/src/runs/StepLogsDialog.tsx
@@ -17,7 +17,7 @@ import {CapturedOrExternalLogPanel} from './CapturedLogPanel';
 import {ComputeLogPanel} from './ComputeLogPanel';
 import {DefaultLogLevels} from './LogLevel';
 import {LogFilter, LogsProvider, LogsProviderLogs} from './LogsProvider';
-import {LogsScrollingTable} from './LogsScrollingTable';
+import {LogsScrollingTableNew} from './LogsScrollingTableNew';
 import {LogType, LogsToolbar} from './LogsToolbar';
 import {IRunMetadataDict, RunMetadataProvider} from './RunMetadataProvider';
 import {titleForRun} from './RunUtils';
@@ -182,7 +182,7 @@ export const StepLogsModalContent = ({
           />
         )
       ) : (
-        <LogsScrollingTable
+        <LogsScrollingTableNew
           logs={logs}
           filter={filter}
           filterStepKeys={stepKeys}
